+++ sketch/src/main/java/com/github/panpf/sketch/stateimage/ErrorStateImage.kt
@@ -16,12 +16,10 @@
 package com.github.panpf.sketch.stateimage
 
 import android.graphics.drawable.Drawable
-import com.github.panpf.sketch.Sketch
 import com.github.panpf.sketch.request.ImageRequest
 import com.github.panpf.sketch.request.UriInvalidException
 import com.github.panpf.sketch.stateimage.ErrorStateImage.Builder
-import com.github.panpf.sketch.stateimage.ErrorStateImage.ErrorRules
-import com.github.panpf.sketch.util.ifOrNull
+import com.github.panpf.sketch.stateimage.internal.CompositeStateImage
 import java.util.LinkedList
 
 /**
@@ -35,142 +33,86 @@ fun ErrorStateImage(
 }.build()
 
 /**
- * Provide Drawable specifically for error status, support custom [ErrorRules] Provide different Drawable according to different error types
+ * Provide Drawable specifically for error status, support custom [CompositeStateImage.Condition] Provide different Drawable according to different error types
  */
-interface ErrorStateImage : StateImage {
+class ErrorStateImage(override val stateList: List<Pair<CompositeStateImage.Condition, StateImage?>>) :
+    CompositeStateImage {
 
-    val errorRulesList: List<ErrorRules>
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (javaClass != other?.javaClass) return false
+        other as ErrorStateImage
+        if (stateList != other.stateList) return false
+        return true
+    }
 
-    override fun getDrawable(
-        sketch: Sketch,
-        request: ImageRequest,
-        throwable: Throwable?
-    ): Drawable? = errorRulesList
-        .firstNotNullOfOrNull { it.getDrawable(sketch, request, throwable) }
-        ?.getOrNull()
+    override fun hashCode(): Int {
+        return stateList.hashCode()
+    }
+
+    override fun toString(): String {
+        val listString = stateList.joinToString(prefix = "[", postfix = "]")
+        return "ErrorStateImage(${listString})"
+    }
 
     class Builder constructor(private val defaultImage: StateImage?) {
 
-        private val errorRulesList = LinkedList<ErrorRules>()
+        private val stateList = LinkedList<Pair<CompositeStateImage.Condition, StateImage?>>()
 
         /**
-         * Add a custom [ErrorRules]
+         * Add a custom state
          */
-        fun addErrorRules(errorRules: ErrorRules): Builder = apply {
-            errorRulesList.add(errorRules)
+        fun addState(pair: Pair<CompositeStateImage.Condition, StateImage?>): Builder = apply {
+            stateList.add(pair)
         }
 
         /**
          * Add a StateImage dedicated to the empty uri error
          */
         fun uriEmptyError(emptyImage: StateImage): Builder = apply {
-            addErrorRules(UriEmptyErrorRules(emptyImage))
+            addState(UriEmptyCondition to emptyImage)
         }
 
         /**
          * Add a StateImage dedicated to the empty uri error
          */
         fun uriEmptyError(emptyDrawable: Drawable): Builder = apply {
-            addErrorRules(UriEmptyErrorRules(DrawableStateImage(emptyDrawable)))
+            addState(UriEmptyCondition to DrawableStateImage(emptyDrawable))
         }
 
         /**
          * Add a StateImage dedicated to the empty uri error
          */
         fun uriEmptyError(emptyImageResId: Int): Builder = apply {
-            addErrorRules(UriEmptyErrorRules(DrawableStateImage(emptyImageResId)))
+            addState(UriEmptyCondition to DrawableStateImage(emptyImageResId))
         }
 
         fun build(): ErrorStateImage {
             val list = if (defaultImage != null) {
-                errorRulesList.plus(DefaultErrorRules(defaultImage))
+                stateList.plus(DefaultCondition to defaultImage)
             } else {
-                errorRulesList
+                stateList
             }
-            return ErrorStateImageImpl(list)
+            return ErrorStateImage(list)
         }
     }
 
-    class ErrorStateImageImpl(override val errorRulesList: List<ErrorRules>) :
-        ErrorStateImage {
+    object DefaultCondition : CompositeStateImage.Condition {
 
-        override fun equals(other: Any?): Boolean {
-            if (this === other) return true
-            if (javaClass != other?.javaClass) return false
-            other as ErrorStateImageImpl
-            if (errorRulesList != other.errorRulesList) return false
-            return true
-        }
-
-        override fun hashCode(): Int {
-            return errorRulesList.hashCode()
-        }
+        override fun accept(request: ImageRequest, throwable: Throwable?): Boolean = true
 
         override fun toString(): String {
-            val factoryListString =
-                errorRulesList.joinToString(prefix = "[", postfix = "]")
-            return "ErrorStateImage(${factoryListString})"
+            return "DefaultCondition"
         }
     }
 
-    /**
-     * Match the error and return a dedicated Drawable
-     */
-    interface ErrorRules {
+    object UriEmptyCondition : CompositeStateImage.Condition {
 
-        fun getDrawable(
-            sketch: Sketch,
-            request: ImageRequest,
-            throwable: Throwable?
-        ): Result<Drawable?>?
-    }
-
-    class DefaultErrorRules(private val stateImage: StateImage) : ErrorRules {
-
-        override fun getDrawable(
-            sketch: Sketch, request: ImageRequest, throwable: Throwable?
-        ): Result<Drawable?> = Result.success(stateImage.getDrawable(sketch, request, throwable))
-
-        override fun equals(other: Any?): Boolean {
-            if (this === other) return true
-            if (javaClass != other?.javaClass) return false
-            other as DefaultErrorRules
-            if (stateImage != other.stateImage) return false
-            return true
-        }
-
-        override fun hashCode(): Int {
-            return stateImage.hashCode()
-        }
+        override fun accept(request: ImageRequest, throwable: Throwable?): Boolean =
+            throwable is UriInvalidException && (request.uriString.isEmpty() || request.uriString.isBlank())
 
         override fun toString(): String {
-            return "DefaultErrorRules($stateImage)"
-        }
-    }
-
-    class UriEmptyErrorRules(private val stateImage: StateImage) : ErrorRules {
-
-        override fun getDrawable(
-            sketch: Sketch, request: ImageRequest, throwable: Throwable?
-        ): Result<Drawable?>? =
-            ifOrNull(throwable is UriInvalidException && (request.uriString.isEmpty() || request.uriString.isBlank())) {
-                Result.success(stateImage.getDrawable(sketch, request, throwable))
-            }
-
-        override fun equals(other: Any?): Boolean {
-            if (this === other) return true
-            if (javaClass != other?.javaClass) return false
-            other as UriEmptyErrorRules
-            if (stateImage != other.stateImage) return false
-            return true
-        }
-
-        override fun hashCode(): Int {
-            return stateImage.hashCode()
-        }
-
-        override fun toString(): String {
-            return "UriEmptyErrorRules($stateImage)"
+            return "UriEmptyCondition"
         }
     }
 }